split the program into files
