+++ Task0/Words_reader.cpp
@@ -1,0 +1,54 @@
+#include <string>
+#include <string>
+#include <vector>
+#include <map>
+#include <fstream>
+#include "Words_reader.h"
+bool is_separator (const char &symbol) {
+    return (!(std::isdigit(symbol) || std::isalpha(symbol)));
+}
+
+std::map<std::string, int32_t> Words_classification (std::vector<std::string> words) {
+    std::map<std::string, int32_t> frequency;
+    for (std:: size_t i = 0; i < words.size(); ++i) {
+        auto it = frequency.find(words[i]);
+        if (it != frequency.end()) {
+            it->second +=1;
+        }
+        else {
+            frequency[words[i]] = 1;
+        }
+    }
+    return frequency;
+}
+std::vector<std::string> Get_lines (std::string input) {
+    std::ifstream input_file (input);
+    std::vector<std::string> lines;
+    std::string line;
+    while (std::getline(input_file, line)) {
+        lines.push_back(line);
+    }
+    return lines;
+}
+
+std::vector<std::string> Parse_into_words (const std::string &line) {
+    char symbol = 0;
+    std::string word = "";
+    std::vector<std::string> words = std::vector<std::string>();
+    for (std:: size_t i = 0; i < line.length(); ++i) {
+        symbol = line[i];
+        if (!is_separator(symbol)) {
+            word = word + symbol;
+            if (i == line.length() - 1) {
+                words.push_back(word);
+            }
+        }
+        else {
+            if (word !="") {
+                words.push_back(word);
+            }
+            word = "";
+        }
+    }
+    return words;
+}
